+++ packages/react-editor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @chamaeleon/react-editor
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @chamaeleon/core@0.6.0
+
 ## 0.0.9
 
 ### Patch Changes